feat: 完成“添加管理员”接口
- 新增 AdminInsertDTO 类
- 新增 AdminDetailVO 类
- 全局异常处理器添加处理 MethodArgumentNotValidException 异常方法

+++ src/main/java/cn/edu/gdufs/pojo/Admin.java
@@ -43,7 +43,7 @@ public class Admin {
    * 管理员权限：详见 RoleConstant 类
    */
   @Min(value = 0,message = "管理员权限不能小于0")
-  private Long role;
+  private Integer role;
 
   /**
    * 管理员昵称
